# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the devolt api server to manage devices.",
        "title": "Devices Api Server",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "DeVolt Team",
            "url": "https://devolt.xyz",
            "email": "henrique@mugen.builders"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
        "/device": {
            "get": {
                "description": "GET all devices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Find All Devices",
                "operationId": "find-all-devices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.FindAllDevicesOutputDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "Create Device with given latitude, longitude, and params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create Device",
                "operationId": "create-device",
                "parameters": [
                    {
                        "description": "Device data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateDeviceInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device created successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateDeviceInputDTO": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "params": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
        "dto.FindAllDevicesOutputDTO": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "params": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        }
    }
}